requirements.txt
prefect>=3.0
pandas>=2.0
beautifulsoup4>=4.12
requests>=2.31
sqlalchemy>=2.0
psycopg2-binary>=2.9
streamlit>=1.30
plotly>=5.18
python-dotenv>=1.0
